Split bridge deposit token tests
