+++ .pm/tasks/task_b80b6013f9a74f78b4dae611f5a408b9.execution.md
@@ -225,3 +225,19 @@ Example:
 - PR Purpose Decision: `normal_pr_ci_watch`
 - Rationale: ordinary implementation/documentation PR, not a manual packaging/release CI hold.
 - Next Watch Gates: GitHub required checks, mergeability, requested changes, PR comments, and unresolved review threads.
+
+## 2026-06-10 22:33:42 CST / tpm
+- CI Failure: PR #397 `required-gate` failed in `./scripts/check-rust-file-size.sh`.
+- Failure Signature: `crates/oasis7/src/bin/oasis7_newapi_bridge_service/tests.rs` had 1313 lines; gate limit is 1200.
+- Root Cause: deposit-token coverage added enough test code to push the existing bridge service test module over the repository file-size limit.
+- Fix: moved deposit-token-specific route/reconcile tests into `crates/oasis7/src/bin/oasis7_newapi_bridge_service/deposit_token_tests.rs` and included it as a child test module. No service behavior changed.
+- Validation Command: `wc -l crates/oasis7/src/bin/oasis7_newapi_bridge_service/tests.rs crates/oasis7/src/bin/oasis7_newapi_bridge_service/deposit_token_tests.rs`
+- Actual Result: `tests.rs` 1081 lines; `deposit_token_tests.rs` 235 lines.
+- Validation Command: `env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_newapi_bridge_service -- --nocapture --test-threads=1`
+- Actual Result: Passed; 31 tests passed, 0 failed. The deposit-token tests were discovered under `tests::deposit_token_tests::*`.
+- Validation Command: `env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_chain_transfer_submit_client -- --nocapture`
+- Actual Result: Passed; 6 tests passed, 0 failed.
+- Validation Command: `env -u RUSTC_WRAPPER cargo fmt --check && git diff --check && ./scripts/check-rust-file-size.sh`
+- Actual Result: Passed; file-size gate reports oversized code files=0, test files=0, structural slice files=0, include targets=0.
+- Validation Command: `./scripts/doc-governance-check.sh`
+- Actual Result: Passed: `doc-governance-check: OK`.
